requirements.txt
pandas
numpy
matplotlib
PyYAML
scikit-learn
# -e .